+++ sdk/planetarycomputer/Azure.ResourceManager.PlanetaryComputer/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.1.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,20 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0 (2026-04-30)
+
+### Other Changes
+
+- This is the first stable release of the `Azure.ResourceManager.PlanetaryComputer` library.
+
+## 1.0.0-beta.2 (2026-04-28)
+
+### Other Changes
+
+- Upgraded API version to 2026-04-15.
+- Upgraded dependent `Azure.Core` to 1.54.0.
+- Upgraded dependent `Azure.ResourceManager` to 1.14.0.
 
 ## 1.0.0-beta.1 (2025-07-11)